* Fixed #658: hosting plan can be deleted but there is another order with it * Fixed #659: resulted of ticket #658 an order cant be update if the hosting plan is deleted * Updated French translation (thanks to Grégory Marigot)
git-svn-id: https://i-mscp.svn.sourceforge.net/svnroot/i-mscp/trunk@812 15a38060-5cfe-4f5c-bf9d-b5d42b1378a7

+++ gui/reseller/dhp.php
@@ -1,21 +1,21 @@
 <?php
 /**
- *  ispCP (OMEGA) - Virtual Hosting Control System | Omega Version
+ * ispCP ω (OMEGA) a Virtual Hosting Control System
  *
- *  @copyright 	2001-2006 by moleSoftware GmbH
- *  @copyright 	2006-2007 by ispCP | http://isp-control.net
- *  @link 		http://isp-control.net
- *  @author		ispCP Team (2007)
+ * @copyright 	2001-2006 by moleSoftware GmbH
+ * @copyright 	2006-2007 by ispCP | http://isp-control.net
+ * @link 		http://isp-control.net
+ * @author 		ispCP Team (2007)
  *
- *  @license
- *  This program is free software; you can redistribute it and/or modify it under
- *  the terms of the MPL General Public License as published by the Free Software
- *  Foundation; either version 1.1 of the License, or (at your option) any later
- *  version.
- *  You should have received a copy of the MPL Mozilla Public License along with
- *  this program; if not, write to the Open Source Initiative (OSI)
- *  http://opensource.org | osi@opensource.org
- **/
+ * @license
+ *   This program is free software; you can redistribute it and/or modify it under
+ *   the terms of the MPL General Public License as published by the Free Software
+ *   Foundation; either version 1.1 of the License, or (at your option) any later
+ *   version.
+ *   You should have received a copy of the MPL Mozilla Public License along with
+ *   this program; if not, write to the Open Source Initiative (OSI)
+ *   http://opensource.org | osi@opensource.org
+ */
 
 require '../include/ispcp-lib.php';
 
@@ -23,22 +23,32 @@ check_login(__FILE__);
 
 $theme_color = $cfg['USER_INITIAL_THEME'];
 
-if(isset($_GET['hpid']) && is_numeric($_GET['hpid']))
+if (isset($_GET['hpid']) && is_numeric($_GET['hpid']))
 	$hpid = $_GET['hpid'];
-else{
+else {
 	$_SESSION['hp_deleted'] = '_no_';
 	Header("Location: hp.php");
 	die();
 }
 
+// Check if there is no order for this plan
+$res = exec_query($sql, "SELECT COUNT(id) FROM orders WHERE plan_id=?", array($hpid));
+$data = $res->FetchRow();
+if ($data['0'] > 0) {
+	$_SESSION['hp_deleted_ordererror'] = '_yes_';
+	header("Location: hp.php");
+	die();
+}
+
+
 // Try to delete hosting plan from db
-//
+
 $query = "delete from hosting_plans where id=? and reseller_id=?";
-$res = exec_query($sql, $query, array($hpid,$_SESSION['user_id']));
+$res = exec_query($sql, $query, array($hpid, $_SESSION['user_id']));
 
 $_SESSION['hp_deleted'] = '_yes_';
 
-Header("Location: hp.php");
+header("Location: hp.php");
 
 die();
